third_party/blink/web_tests/external/wpt/editing/run/removeformat-expected.txt

This is a testharness.js-based test.
Found 45 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] [["removeformat",""]] "foo[]bar": execCommand("removeformat", false, "") return value
  assert_equals: expected true but got false
[FAIL] [["removeformat",""]] "<span>foo</span>{}<span>bar</span>": execCommand("removeformat", false, "") return value
  assert_equals: expected true but got false
[FAIL] [["removeformat",""]] "<span>foo[</span><span>]bar</span>": execCommand("removeformat", false, "") return value
  assert_equals: expected true but got false
[FAIL] [["stylewithcss","true"],["removeformat",""]] "[foo<b>bar</b>baz]" queryCommandState("stylewithcss") before
  assert_equals: Wrong result returned expected false but got true
[FAIL] [["stylewithcss","true"],["removeformat",""]] "[foo<span style=\\"display: none\\">bar</span>baz]" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foobarbaz" but got "foo<span style=\\"display:none\\">bar</span>baz"
[FAIL] [["stylewithcss","false"],["removeformat",""]] "[foo<span style=\\"display: none\\">bar</span>baz]" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foobarbaz" but got "foo<span style=\\"display:none\\">bar</span>baz"
[FAIL] [["stylewithcss","true"],["removeformat",""]] "[foo<span style=\\"display: none; font-weight: bold\\">bar</span>baz]" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foobarbaz" but got "foo<span style=\\"display:none\\">bar</span>baz"
[FAIL] [["stylewithcss","false"],["removeformat",""]] "[foo<span style=\\"display: none; font-weight: bold\\">bar</span>baz]" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foobarbaz" but got "foo<span style=\\"display:none\\">bar</span>baz"
[FAIL] [["stylewithcss","true"],["removeformat",""]] "foo<span style=\\"display: none\\">b[a]r</span>baz": execCommand("removeformat", false, "") return value
  assert_equals: expected true but got false
[FAIL] [["stylewithcss","true"],["removeformat",""]] "foo<span style=\\"display: none\\">b[a]r</span>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<span style=\\"display:none\\">b</span>a<span style=\\"display:none\\">r</span>baz" but got "foo<span style=\\"display:none\\">bar</span>baz"
[FAIL] [["stylewithcss","false"],["removeformat",""]] "foo<span style=\\"display: none\\">b[a]r</span>baz": execCommand("removeformat", false, "") return value
  assert_equals: expected true but got false
[FAIL] [["stylewithcss","false"],["removeformat",""]] "foo<span style=\\"display: none\\">b[a]r</span>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<span style=\\"display:none\\">b</span>a<span style=\\"display:none\\">r</span>baz" but got "foo<span style=\\"display:none\\">bar</span>baz"
[FAIL] [["stylewithcss","true"],["removeformat",""]] "foo<span style=\\"display: none; font-weight: bold\\">b[a]r</span>baz": execCommand("removeformat", false, "") return value
  assert_equals: expected true but got false
[FAIL] [["stylewithcss","true"],["removeformat",""]] "foo<span style=\\"display: none; font-weight: bold\\">b[a]r</span>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<span style=\\"display:none; font-weight:bold\\">b</span>a<span style=\\"display:none; font-weight:bold\\">r</span>baz" but got "foo<span style=\\"display:none; font-weight:bold\\">bar</span>baz"
[FAIL] [["stylewithcss","false"],["removeformat",""]] "foo<span style=\\"display: none; font-weight: bold\\">b[a]r</span>baz": execCommand("removeformat", false, "") return value
  assert_equals: expected true but got false
[FAIL] [["stylewithcss","false"],["removeformat",""]] "foo<span style=\\"display: none; font-weight: bold\\">b[a]r</span>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<span style=\\"display:none; font-weight:bold\\">b</span>a<span style=\\"display:none; font-weight:bold\\">r</span>baz" but got "foo<span style=\\"display:none; font-weight:bold\\">bar</span>baz"
[FAIL] [["stylewithcss","true"],["removeformat",""]] "[foo<span style=\\"font-variant: small-caps\\">bar</span>baz]" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foobarbaz" but got "foo<span style=\\"font-variant-numeric:normal; font-variant-east-asian:normal; font-variant-alternates:normal; font-variant-position:normal; font-variant-emoji:normal\\">bar</span>baz"
[FAIL] [["stylewithcss","false"],["removeformat",""]] "[foo<span style=\\"font-variant: small-caps\\">bar</span>baz]" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foobarbaz" but got "foo<span style=\\"font-variant-numeric:normal; font-variant-east-asian:normal; font-variant-alternates:normal; font-variant-position:normal; font-variant-emoji:normal\\">bar</span>baz"
[FAIL] [["stylewithcss","true"],["removeformat",""]] "foo<span style=\\"font-variant: small-caps\\">b[a]r</span>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<span style=\\"font-variant:small-caps\\">b</span>a<span style=\\"font-variant:small-caps\\">r</span>baz" but got "foo<span style=\\"font-variant:small-caps\\">b</span><span style=\\"font-variant-numeric:normal; font-variant-east-asian:normal; font-variant-alternates:normal; font-variant-position:normal; font-variant-emoji:normal\\">a</span><span style=\\"font-variant:small-caps\\">r</span>baz"
[FAIL] [["stylewithcss","false"],["removeformat",""]] "foo<span style=\\"font-variant: small-caps\\">b[a]r</span>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<span style=\\"font-variant:small-caps\\">b</span>a<span style=\\"font-variant:small-caps\\">r</span>baz" but got "foo<span style=\\"font-variant:small-caps\\">b</span><span style=\\"font-variant-numeric:normal; font-variant-east-asian:normal; font-variant-alternates:normal; font-variant-position:normal; font-variant-emoji:normal\\">a</span><span style=\\"font-variant:small-caps\\">r</span>baz"
[FAIL] [["removeformat",""]] "[foo<abbr>bar</abbr>baz]" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foobarbaz" but got "foo<abbr>bar</abbr>baz"
[FAIL] [["removeformat",""]] "foo<abbr>b[a]r</abbr>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<abbr>b</abbr>a<abbr>r</abbr>baz" but got "foo<abbr>bar</abbr>baz"
[FAIL] [["removeformat",""]] "[foo<bdi dir=rtl>bar</bdi>baz]" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foobarbaz" but got "foo<bdi dir=\\"rtl\\">bar</bdi>baz"
[FAIL] [["removeformat",""]] "foo<bdi dir=rtl>b[a]r</bdi>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<bdi dir=\\"rtl\\">b</bdi>a<bdi dir=\\"rtl\\">r</bdi>baz" but got "foo<bdi dir=\\"rtl\\">bar</bdi>baz"
[FAIL] [["removeformat",""]] "[foo<blink>bar</blink>baz]" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foobarbaz" but got "foo<blink>bar</blink>baz"
[FAIL] [["removeformat",""]] "foo<blink>b[a]r</blink>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<blink>b</blink>a<blink>r</blink>baz" but got "foo<blink>bar</blink>baz"
[FAIL] [["removeformat",""]] "[foo<mark>bar</mark>baz]" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foobarbaz" but got "foo<mark>bar</mark>baz"
[FAIL] [["removeformat",""]] "foo<mark>b[a]r</mark>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<mark>b</mark>a<mark>r</mark>baz" but got "foo<mark>bar</mark>baz"
[FAIL] [["removeformat",""]] "[foo<span>bar</span>baz]" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foobarbaz" but got "foo<span>bar</span>baz"
[FAIL] [["removeformat",""]] "foo<span>b[a]r</span>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<span>b</span>a<span>r</span>baz" but got "foobarbaz"
[FAIL] [["removeformat",""]] "[foo<span id=foo>bar</span>baz]" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foobarbaz" but got "foo<span id=\\"foo\\">bar</span>baz"
[FAIL] [["removeformat",""]] "foo<span id=foo>b[a]r</span>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<span id=\\"foo\\">b</span>a<span>r</span>baz" but got "foo<span id=\\"foo\\">bar</span>baz"
[FAIL] [["removeformat",""]] "[foo<span class=foo>bar</span>baz]" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foobarbaz" but got "foo<span class=\\"foo\\">bar</span>baz"
[FAIL] [["removeformat",""]] "foo<span class=foo>b[a]r</span>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<span class=\\"foo\\">b</span>a<span class=\\"foo\\">r</span>baz" but got "foo<span class=\\"foo\\">bar</span>baz"
[FAIL] [["removeformat",""]] "<p style=\\"background-color: aqua\\">foo[bar]baz</p>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p style=\\"background-color:rgb(0, 255, 255)\\">foobarbaz</p>" but got "<p style=\\"\\"><span style=\\"background-color:rgb(0, 255, 255)\\">foo</span>bar<span style=\\"background-color:rgb(0, 255, 255)\\">baz</span></p>"
[FAIL] [["stylewithcss","true"],["removeformat",""]] "<p style=\\"font-weight: bold\\">foo[bar]baz</p>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><span style=\\"font-weight:bold\\">foo</span>bar<span style=\\"font-weight:bold\\">baz</span></p>" but got "<p style=\\"\\"><span style=\\"font-weight:bold\\">foo</span>bar<span style=\\"font-weight:bold\\">baz</span></p>"
[FAIL] [["stylewithcss","false"],["removeformat",""]] "<p style=\\"font-weight: bold\\">foo[bar]baz</p>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><b>foo</b>bar<b>baz</b></p>" but got "<p style=\\"\\"><span style=\\"font-weight:bold\\">foo</span>bar<b>baz</b></p>"
[FAIL] [["stylewithcss","true"],["removeformat",""]] "<b><p style=\\"font-weight: bold\\">foo[bar]baz</p></b>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><span style=\\"font-weight:bold\\">foo</span>bar<span style=\\"font-weight:bold\\">baz</span></p>" but got "<p style=\\"\\"><b><span style=\\"font-weight:bold\\">foo</span></b>bar<b>baz</b></p>"
[FAIL] [["stylewithcss","false"],["removeformat",""]] "<b><p style=\\"font-weight: bold\\">foo[bar]baz</p></b>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><b>foo</b>bar<b>baz</b></p>" but got "<p style=\\"\\"><b><span style=\\"font-weight:bold\\">foo</span></b>bar<b>baz</b></p>"
[FAIL] [["removeformat",""]] "<p style=\\"font-variant: small-caps\\">foo[bar]baz</p>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p style=\\"font-variant:small-caps\\">foobarbaz</p>" but got "<p style=\\"font-variant-numeric:normal; font-variant-east-asian:normal; font-variant-alternates:normal; font-variant-position:normal; font-variant-emoji:normal\\"><span style=\\"font-variant-ligatures:normal; font-variant-caps:small-caps\\">foo</span>bar<span style=\\"font-variant-caps:small-caps\\">baz</span></p>"
[FAIL] [["removeformat",""]] "{<p style=\\"font-variant: small-caps\\">foobarbaz</p>}" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p style=\\"font-variant:small-caps\\">foobarbaz</p>" but got "<p style=\\"font-variant-numeric:normal; font-variant-east-asian:normal; font-variant-alternates:normal; font-variant-position:normal; font-variant-emoji:normal\\">foobarbaz</p>"
[FAIL] [["removeformat",""]] "<p style=\\"text-indent: 2em\\">foo[bar]baz</p>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p style=\\"text-indent:2em\\">foobarbaz</p>" but got "<p style=\\"\\">foobarbaz</p>"
[FAIL] [["removeformat",""]] "{<p style=\\"text-indent: 2em\\">foobarbaz</p>}" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p style=\\"text-indent:2em\\">foobarbaz</p>" but got "<p style=\\"\\">foobarbaz</p>"
[FAIL] [["stylewithcss","true"],["removeformat",""]] "<table data-start=0 data-end=1><tr><td><b>foo</b></table>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<table><tbody><tr><td>foo</td></tr></tbody></table>" but got "<table><tbody><tr><td><b>foo</b></td></tr></tbody></table>"
[FAIL] [["stylewithcss","false"],["removeformat",""]] "<table data-start=0 data-end=1><tr><td><b>foo</b></table>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<table><tbody><tr><td>foo</td></tr></tbody></table>" but got "<table><tbody><tr><td><b>foo</b></td></tr></tbody></table>"
Harness: the test ran to completion.
